+++ src/error.rs
@@ -148,6 +148,15 @@ impl From<ort::Error> for TransmutationError {
     }
 }
 
+impl From<zip::result::ZipError> for TransmutationError {
+    fn from(err: zip::result::ZipError) -> Self {
+        TransmutationError::IoError(std::io::Error::new(
+            std::io::ErrorKind::Other,
+            err.to_string()
+        ))
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;